+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/SlimefunPlugin.java
@@ -41,7 +41,7 @@ import io.github.thebusybiscuit.slimefun4.core.services.PerWorldSettingsService;
 import io.github.thebusybiscuit.slimefun4.core.services.PermissionsService;
 import io.github.thebusybiscuit.slimefun4.core.services.UpdaterService;
 import io.github.thebusybiscuit.slimefun4.core.services.github.GitHubService;
-import io.github.thebusybiscuit.slimefun4.core.services.metrics.MetricsService;
+import io.github.thebusybiscuit.slimefun4.core.services.MetricsService;
 import io.github.thebusybiscuit.slimefun4.core.services.plugins.ThirdPartyPluginService;
 import io.github.thebusybiscuit.slimefun4.core.services.profiler.SlimefunProfiler;
 import io.github.thebusybiscuit.slimefun4.implementation.items.altar.AncientAltar;
@@ -196,7 +196,7 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
             networkManager = new NetworkManager(networkSize);
 
             // Setting up bStats
-            metricsService.start();
+            new Thread(metricsService::start).start();
 
             // Starting the Auto-Updater
             if (config.getBoolean("options.auto-update")) {
@@ -366,6 +366,8 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
 
         // Create a new backup zip
         backupService.run();
+
+        metricsService.cleanUp();
 
         // Prevent Memory Leaks
         // These static Maps should be removed at some point...
@@ -562,6 +564,16 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
      */
     public static UpdaterService getUpdater() {
         return instance.updaterService;
+    }
+
+    /**
+     * This method returns the {@link MetricsService} of Slimefun.
+     * It is used to handle sending metric information to bStats.
+     *
+     * @return The {@link MetricsService} for Slimefun
+     */
+    public static MetricsService getMetricsService() {
+        return instance.metricsService;
     }
 
     /**